+++ app/opulence/orm/datamappers/MemcachedCachedSQLDataMapper.php
@@ -6,16 +6,16 @@
  */
 namespace Opulence\ORM\DataMappers;
 
-use Opulence\Memcached\OpulenceMemcached;
 use Opulence\Databases\ConnectionPool;
+use Opulence\Memcached\Memcached;
 
 abstract class MemcachedCachedSQLDataMapper extends CachedSQLDataMapper
 {
     /**
-     * @param OpulenceMemcached $cache The cache object used in the cache data mapper
+     * @param Memcached $cache The cache object used in the cache data mapper
      * @param ConnectionPool $connectionPool The connection pool used in the SQL data mapper
      */
-    public function __construct(OpulenceMemcached $cache, ConnectionPool $connectionPool)
+    public function __construct(Memcached $cache, ConnectionPool $connectionPool)
     {
         parent::__construct($cache, $connectionPool);
     }