+++ app/opulence/orm/datamappers/RedisCachedSQLDataMapper.php
@@ -6,16 +6,16 @@
  */
 namespace Opulence\ORM\DataMappers;
 
-use Opulence\Redis\IRedis;
 use Opulence\Databases\ConnectionPool;
+use Opulence\Redis\Redis;
 
 abstract class RedisCachedSQLDataMapper extends CachedSQLDataMapper
 {
     /**
-     * @param IRedis $cache The cache object used in the cache data mapper
+     * @param Redis $cache The cache object used in the cache data mapper
      * @param ConnectionPool $connectionPool The connection pool used in the SQL data mapper
      */
-    public function __construct(IRedis $cache, ConnectionPool $connectionPool)
+    public function __construct(Redis $cache, ConnectionPool $connectionPool)
     {
         parent::__construct($cache, $connectionPool);
     }